requirements.txt
fabric
dj-database-url
python-decouple